backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Request handling
python-multipart==0.0.6

# LangChain and LLM integration
langchain==0.1.0
langchain-openai==0.0.5

# Vector database
chromadb==0.4.15

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-markdown>=3.5.1

# Database ORM and migrations
sqlalchemy==2.0.23
alembic==1.12.1

# Testing
pytest==7.4.0
httpx==0.25.0
